Disallow pip from fetching from PyPI.
This wasn't happening for anything we ship because all of our
requirements are already local paths, but for some reason pip tries to
update setuptools for itself with 3.11, and it does that by going to
pypi. That's failing because the python prebuilt doesn't support SSL,
but we don't want it doing that anyway.

Bug: None
Test: None
Change-Id: I6a904e10cdccddc5503a20fcbe906cf592c7465f