Preserve symlinks when copying to the app bundle.
We'd done all the work to set these up and make sure they're preserved
when zipping, but missed the preservation flag on copy :(

This was caught because the app bundle had finally crossed the 4GiB
threshold for zip32, which broke the signing process. This will get us
back down to 3GiB ish, which buys us time, even if it's not a permanent
fix.

Bug: http://b/333446731
Test: ./checkbuild.py --package && du -sh ../out/bundle
Change-Id: Ic22a2c0e0a4a8a24cc639a7a72edb19a92d0dc67