Add install path info to the source.properties.
ADRT used to have some explicit mangling of the Pkg.Revision to remove
this extra data. I have no idea why that was done, as it makes the
install path worse, but we have to remain compatible with it. The new
release tooling doesn't do that mangling (nor should it!), so set the
install path explicitly here.

Bug: http://b/333988131#comment7
Test: inspected source.properties after build
Change-Id: Iedd27a0d275d50f0ee312f1c7a8587a7f761e153