Update the roadmap.
Apparently I've forgotten to update this for a while. A handful of
things get to be moved to the "done" section (hooray!), but a number of
others move from "current" to "future" as we haven't had the bandwidth
to pursue them (boo). One thing does get to move from "future" to
"current" though (improving samples), so that's nice.

Bug: None
Test: None
Change-Id: Ibd8ae3e47d55f90eb99d34f2d116ed834e54ae67