Reformat changelog with prettier.
I'm getting tired of disabling my autoformatter when I touch these.
Annoyingly prettier doesn't have an option for whether to use * or - for
lists (and if you dig a bit, you'll find that they've flipped back and
forth over time), but I don't really care which is used if I don't have
to do it by hand.

Bug: None
Test: None
Change-Id: Ic53b2e7c60470e2872e318d4d9c5ba557adcce89