Remove toolchain runtimes for non-Android targets.
We don't support building host code from the NDK (and anyone that does
that regardless probably won't be using these anyway since they'll be
targeting a GNU system that uses libgcc), but the platform build does,
so there's a bunch of extra target runtime libraries in the toolchain
directory. Purge them to save disk space.

Bug: None
Test: treehugger to check that nothing broke, ls to check files are gone
Change-Id: I8c2b71e901552f521f198123ba855a3de9f164b1