Fix CMAKE_SYSTEM_PROCESSOR CMake use case.
This isn't actually the documented interface for CMake without a
toolchain file, but a user reported it as broken for them. It's trivial
to put the data back, so why not.

I've added a test, but it's pretty useless. The built library won't ever
make it to a device because it's a build-only test. Adding full test
coverage for this workflow would be a fair amount more work, and because
we *expect* that workflow to have behavior differences (if it didn't, we
wouldn't need the toolchain file), it'd always be a case where we were
disabling tests as not supported.

Still, we should make sure we can at least configure a build correctly
in this mode. Anything else that goes wrong is almost certainly the
result of avoiding all the other stuff the toolchain file does, so WAI.

The test actually would not have caught the breakage the user reported,
because the break is only present in specifically the
CMAKE_SYSTEM_PROCESSOR case; skipping the toolchain file but using the
documented CMAKE_ANDROID_ARCH_ABI option instead still worked.

Bug: https://github.com/android/ndk/issues/2049
Test: manual edit of the added test case to make sure I fixed this, but
      reverted that change because it's not actually the documented
      CMake interface
(cherry picked from https://android-review.googlesource.com/q/commit:b49fc7f30d84241dba753d952e4531fb503bbf33)
Merged-In: I614dab1c72a9ee3814e050393dac8c9adc677f51
Change-Id: I614dab1c72a9ee3814e050393dac8c9adc677f51