Clear LOCAL_STRIP_MODE in CLEAR_VARS.
There isn't a very good way to write a useful test for this. I can write
a regression test for exactly the `STRIP_MODE` case, but all that does
is verify that `STRIP_MODE` hasn't disappeared from the list, and the
only reason to ever remove something from this list is when an option is
removed. The useful test would be testing that all `LOCAL_*` variables
that are consumed are in this list, but this list only exists because
there is no way to enumerate all the `LOCAL_*` variables that are used,
so there's no way to write that test so long as this is implemented in
a language without variable scopes.

Bug: https://github.com/android/ndk/issues/2039
Test: local test case showing that the strip mode is cleared
(cherry picked from https://android-review.googlesource.com/q/commit:180aeab317e9b4ceffc277e15525626f9aa092eb)
Merged-In: I4d56c64aa792dcc276443d131beeaf329a9a33a5
Change-Id: I4d56c64aa792dcc276443d131beeaf329a9a33a5