Add test-release tool.
This is like run_tests.py but operates on CI artifacts. It automates the
manual steps of downloading the CI-build test artifacts for each host
platform, extracting them, and running run_tests.py with them.

The output is pretty ugly, but that's something that's most easily
cleaned up when I fix the TODO here and reorganize run_tests.py to be
used without shelling out to it.

Bug: None
Test: test-release 12433566
Change-Id: Id818bc922dc74cd1db6114a96cde8615935fd6cb