Remove obsolete quirk of test scanning.
The weird indirection here existing to allow scanning for more than one
type of test, which hasn't been needed for the build.sh or test.py type
tests in a very long time (because we scan the dist dir now rather than
the source dir), and the other use was for the libc++ tests, which are
no longer run here.

There's probably a lot of cleanup that could be done here, but for now
I'm only tackling what's in the way of improving the test-release tool.

Bug: None
Test: ./run_tests.py
Change-Id: I468042c6cbc28ad2c1e7006a52dd4708abc98d92