Support finding symbols in Play symbol zip files.
This allows ndk-stack to search a native-debug-symbols.zip file rather
than a directory when it is passed as the symbol source. That is the
file that AGP creates that can be uploaded to Play for symbolized traces
in the Play console (a directory per ABI with $NAME.so.dbg files in it).

This path doesn't yet have the symbol source caching like the directory
source does. I need to extract the caching from the
DirectorySymbolSource and make it its own source which wraps either the
DirectorySymbolSource or the PlayDebugZipSymbolSource. I'll do that in a
follow up.

Bug: https://github.com/android/ndk/issues/2078
Test: Added tests
Change-Id: I824a1f155a027b57fca952c1927d74444412b17c