Raise minimum CMake version.
The latest versions of CMake (3.31 and newer) emit a warning (many times
over the course of a build) whenever the minimum version is specified as
anything older than 3.10. Based on beta feedback, this warning is
intrusive enough that we should break compatibility with users of (very)
old CMake versions.

Studio's stats are reporting the number of builds still on 3.6 (the only
pre-3.10 CMake version in the SDK) as quite low. I unfortunately don't
have that data per NDK version, but intuitively the people sticking with
old version of CMake are probably also sticking with old versions of the
NDK, which should even further reduce the number of users adversely
effected by this. Even if there is anyone effected, upgrading to a new
version of CMake should still be fairly easy for them.

Bug: https://github.com/android/ndk/issues/2100
Test: ./checkbuild.py && ./run_tests.py
Change-Id: Ic5d0be31d05ed87e23d15b9f87bb7b8108d7c227