Begin extracting symbol discovery logic.
This is a weird responsibility for a class called "FrameInfo", and the
organization makes it difficult to add new discovery methods in a way I
can feel confident about. It's also hard to test. This is the first step
in pulling all of `FrameInfo.get_elf_file()` out into a new
`SymbolSource` class, which will have subclasses for each source type.
When I'm done refactoring that'll be the ELF file source, which is the
one I've done here since it's the base for all the other types; an APK
source; and a directory source. Once the refactor is done, I'll add the
new source type for the Play debug symbols zip.

`GetElfFileTests` is going to stick around and be ugly, as will the
`FrameInfo.get_elf_file()` implementation for the time being because
it's even harder to keep the tests running if I rewrite that function in
pieces. Once the refactor is done the tests will be much simpler and
(hopefully) without mocks.

Bug: https://github.com/android/ndk/issues/2078
Test: ./checkbuild.py pytest
Change-Id: Ie7b6cf47b9a5060005808168d16ac9d359d6998e