Unify APK frame handling.
Frames that name APKs might be in the format of APK!library or just the
raw APK name with no library name. Previously a FrameInfo parsed from an
APK-only frame line would lie about what it parsed, claiming that the
APK was the library and there was no container, and the code that
searched for the debug symbols had to catch that lie and correct as
necessary. Doing it that way also meant that we had to check for APK
frames twice. This alters the parsing behavior so that these frames
accurately show that no ELF file was found, and moves the responsibility
for filling that field and rewriting the frame tail into the APK source
handling.

I still don't really like this, but I don't think it can be made much
better. Filling out this information during parse would require
searching and analyzing the APK during parse, but the only reason we're
creating the FrameInfo object in the first place is so we can pass it on
to the scanner.

Bug: https://github.com/android/ndk/issues/2078
Test: added new tests
Change-Id: Ic7c4ccd38591f1a487f575d01eb203c301f9e576