Use symbols as long as the build IDs match.
If the build IDs match, it doesn't matter if the file was renamed.

This also conveniently allows ndk-stack to find symbols in files with
matching build IDs even if all we have is a trace line with no ELF file
name, but no APKs in the symbol directory. Previously we wouldn't have
been able to find that because we wouldn't have known which files to
check without first determining the name of the ELF file by extracting
the APK.

Bug: None
Test: covered by a test
Change-Id: I5a3f7fff160173727c10fc5f708063800d2d6814