Rename BuildIdReader to ElfReader.
The next commit adds debug info detection to the class, so the name is
no longer accurate. The rename's in a separate commit to make it easier
to review.

I've also made the build_id argument of FakeElfReader optional to reduce
clutter in test cases where the build ID doesn't matter.

Bug: https://github.com/android/ndk/issues/2078
Test: pytest doesn't get worse
Change-Id: Ia88e8068903a2e49726223ca6981152d0f3e6dc0