Break up symbolize_trace for testing.
This was pretty hard to test since it created its own dependencies.
This commit breaks it into two classes (TraceSymbolizer and App) so it
can be more easily tested.

I've added some new tests for TraceSymbolizer now that is is testable,
and unsurprisingly found a (minor, formatting only) bug in the process.
That's xfail'd for now and I'll fix it separately.

Bug: None
Test: ./checkbuild.py pythonlint
Change-Id: I06e7b65cae70f8f247254be45b77866b52ad6d7e