Simplify LLVM tools discovery.
This was more complicated than it needs to be in an attempt to be fault
tolerant, but because we can't actually do anything without
llvm-symbolizer, tolerating a missing llvm-readelf doesn't actually help
anyone. Removing that allows simplifying some tests, and makes it much
simpler to extract the tools bin directory dependency for testing, which
the next commit takes advantage of.

Bug: None
Test: ./checkbuild.py pythonlint
Change-Id: Ibacffd02140fcaaf1dd9591c5302db91caab645a