Rework E2E tests to be faster and more correct.
Previously, SystemTests wouldn't catch new issues during development if
you skipped the build process (and of course I do, why wait?), but had
the advantage of testing ndk-stack the way it would actually be used in
the field, which is able to catch packaging bugs.

I've split the test into two forms: one that runs the packaged ndk-stack
"binary" the same as it did before, and a new variant that runs the
program via a Python import, which doesn't require a rebuild to use up-
to-date sources. The latter is much better during development, and the
former is an extra level of security against release bugs.

For easier maintainability, I've reorganized the guts of the tests into
parameterized tests, with a common set of parameters shared between
them.

Bug: None
Test: ./checkbuild.py pythonlint
Change-Id: I7020fd00e247f2a605c886ed9c874d79ed21505c