Lease missing devices from ACID when needed.
If there are tests that require devices that aren't connected to adb, we
can auto-acquire some configurations so the user doesn't have to deal
with it themselves. Right now this is only done via ACID, which can (I
think) only provide x86 and x86_64 emulators, but the same process could
be used to launch Studio emulators which could provide arm/arm64 devices
at least on Arm Macs.

This also enables the Rich logging handler for run_tests.py, which
prevents log messages from breaking the progress UI by printing logs
above the progress bars.

Both of these features rely on things that can't or shouldn't be run on
our build servers (the rich dependency from PyPI, network access for
ACID), but it's constrained to `run_tests.py` which isn't used in CI
anyway.

Test: ./checkbuild.py pythonlint
Test: ./run_tests.py # with no x86/x86_64 devices connected
Test: ./run_tests.py # with x86/x86_64 devices already connected
Test: ./run_tests.py # with no arm devices connected
Change-Id: If53fd2416161336d824fe2432f91942adef92f41