Preserve crtbegin_static when updating the syroot.
Currently we toss all these anyway and compile our own from the bionic
source, but if we preserve these and brand the objects with `ld -r`
rather than compiling our own, we can drop the dependency on the bionic
repo. Losing that dependency should reduce the build rate of the NDK a
lot, since the bionic repo is much more active than the rest of the
NDK's manifest.

Bug: None
Test: updated, crtbegin_static.o is preserved
Change-Id: I1516401c918e66332fedb05c72e44cc5657d010c