Use the CRT objects from the platform.
We've always built our own CRT objects because the platform sysroot
didn't always include them, but they got added to that artifact quite a
while ago. We can brand those objects rather than building our own from
the bionic source, which allows us to drop the dependency on the bionic
repo, and to stop worrying about the build requirements of the CRT
objects.

Bug: None
Test: ./checkbuild.py && ./run_tests.py
Test: ./parse_elfnote.py one/of/the/test/binaries
Change-Id: I8e534b78527b94a619f6ec00eed332df115e08f6