Vendor the gdbrunner source.
I need to break the dependency on //development and there aren't many
good options for that other than forking the source and keeping it in
the NDK.

I'm putting this in sources for now because I don't want to add the full
NDK package to the ndkgdb zipapp, and because this code is currently not
lint-clean. I think I'm going to have to do a big reorganizing of the
NDK package into subpackages so that code that should be shipped to
users, code that should run in CI, and code that's only for local
development environments (the only kind that can have PyPI dependencies)
so it's easy to control what parts of the package ship and which don't.
This will do for now.

Bug: None
Test: ndk-gdb --help
Change-Id: Ia93cb1e6bbd883efa592e83f20ac8c1f53664d3f