Vendor the adb Python source.
Same as the previous commit: breaking the dependency on //development
and this is the easiest way to do it. It can be done more cleanly but
that will require a lot more work inside the ndk package, and a bunch
of new tests.

The alternative solution would be moving this code into its own repo
that's not tied to the rest of the platform. That'd allow me and adb to
continue sharing this Python code, but I'm not sure that's worth it. adb
would rather the adb Python code live inside the main adb repo (and I
agree), but then the NDK would be back to depending on a platform repo
and that's what I'm trying to fix.

On top of that, there's very little code here that'd need to actually be
shared. The NDK needs `adb shell`, `adb sync`, and device enumeration.
The platform uses this as a test harness for adb, so it needs
considerably more, and the platform doesn't need the hacks to work
around bugs and missing features in old versions of adbd like the NDK
does.

I also expect that I'm going to modernize a bunch of this soonish (make
everything async, for one), and don't really want to be on the hook for
keeping all the platform consumers operating when I do that.

Bug: None
Test: ndk-gdb --help
Test: ./run_tests.py
Change-Id: Ia4df4fe4d552b46ea3a362ca3d86d4a40efd4724