Add new CI entry points.
I want to change some workflow that doesn't play nicely with our current
CI setup, and it's annoying to have to go modify a thing in google3
whenever I want to change the CI entry points. This change restructures
that so that the only thing the other end of the CI config needs to do
its own config arguments to ci.sh (or ci.ps1 for Windows), and then our
own code can choose what to do for that build. The Rust toolchain team
has already tried this approach in their build and they seem happy with
it.

Bug: None
Test: ./ci.sh --branch main-ndk --target darwin ...
Change-Id: I298031a5e9d139e00e2a42a8798b46fd7c81ec86