Quarantine and freeze ndk-gdb code.
I want to make some changes to how adb works that aren't needed in
ndk-gdb, and since we don't have tests for ndk-gdb I don't really want
to make changes to the shared component. ndk-gdb isn't getting feature
development any more anyway, so rather than spend a lot of time adding
tests for a thing that's not supposed to be changing, copy the shared
code into the (new) ndk-gdb directory, since that's the closest
equivalent to freezing this non-PyPI dependency.

Bug: None
Test: ndk-gdb.py --help
Change-Id: I448fe13dd646a7fd44cc6255ddcbaabebc277da5