Add new entry point for CI test build.
The --build-only mode got added to run_tests.py a while back for just
this purpose, and aside from the test build target (win64_tests), it's
the only way that run_tests.py gets used outside a local development
environment. Split the rebuild code into its own entry point for CI so
we don't have to worry about CI-specific quirks in run_tests.py. I'll
delete the now obsolete parts once I've seen a green build after this
commit.

This also has the advantage of making the win64_tests path for building
the tests much more like the checkbuild.py path that Darwin and Linux
use. Each is now a thin wrapper for setting up the test build, unlike
the `run_tests.py --build-only` path was. That could probably be
improved slightly more, but it's not likely to be worth the effort.

Change-Id: I140ce1d4e3b1c495d404d70dd403d0a004fb91c0