Remove support for queueing new test builds.
This hasn't been used since we stopped building the libc++ tests. This
was always a crutch for dealing with an unknown number of tests, which
we needed because we delegated test builds to lit and the "additional
tests" here were really just a bunch of results parsed from the XUnit
output. That hack is thankfully no longer needed. Delete it.

Change-Id: I6de80e6329ecb7bfb5cb1c3407a6d4a414b68e1f