Factor out test build UI stuff.
Same thing I just did with the test runner, but now for test builds.

This is probably quite similar to what I'm going to do with the build
itself quite soon, but I've made it a specific test build UI for now
since I'm not sure and will be tackling them separately.

Change-Id: Ibe314570162db5285d7a74784ec79eb38ad7e171