Improve the ndk-build API.
I need to make some changes in the test.py interface that are going to
result in this being used a fair amount more. Make the API better first.

I've also switched it from auto-decoding to str to just return bytes
instead, because this will soon be asyncio, which is a jerk and doesn't
support `encoding` for some reason I'll never understand.

Change-Id: If9df6fbaa33c21da3fdf9050f801daed6dc5203d