Improve the ndk-build API.
I need to make some changes in the test.py interface that are going to
result in this being used a fair amount more. Make the API better first.

I've also switched it from auto-decoding to str to just return bytes
instead, because this will soon be asyncio, which is a jerk and doesn't
support `encoding` for some reason I'll never understand.

Change-Id: Id07d15e47acf0f57f47c34aa9063607342ed818d