Stop printing so much during the build.
This hasn't been a problem before because each worker process would
redirect stdout and stderr to a file. Once those are asyncio tasks
instead of processes, we can't do that per task any more, but we also
won't need to. Capture logs for subprocess exceptions, and use logging
in the few cases where we were printing something potentially useful.

Change-Id: Ie43f051ea935708d0aacf61dd52180e88ae3a18f