Revert to 7 May, build 13461170.
This is the last green build before builds started timing out. They were
flaky for a few builds right before that, so this might still not work,
but I have to revert this far before I can revert more changes on the
go/ab side. I'd be surprised if this works, but I'm out of other things
to try. The release builds are still working and there aren't many
differences between there and here.

This will need to be at least partially reverted once I figure out
what's going on. This is a complete revert for diagnostic purposes, but
because it's complete it's also reverting at least one bug fix.

Change-Id: I7ab6579460f5b8eeafcb704f95640232731a1bfd