Remove global module object list.
Unsurprisingly, this global list of mutable objects causes problems
once the mutation moves from another process into a thread of the main
process, as is happening with the switch to asyncio. This mostly
doesn't cause issues because the mutations are correct throughout a
given host's build, and we only ever build one host at a time. When
cross-compiling for Windows, any *cached* mutations are first
populated for Linux and then not reset for Windows.

This is all kind of gross anyway and the best fix is to just stop
doing it. Remove the global list of instantiated module objects in
favor of just the module types and names (the names are class
properties), and instantiate the objects once per host.

Change-Id: Id620380897961f2e3cf76471ba13b7a72d745680