Log suprocess failures when buidling CRT objects.
I think we used to get this for free because the failures would just
print to stderr, but the new UI causes that to be overwritten. Capturing
the messages like this is better anyway since it prevents them from
interleaving.

It'd be a good idea to go do this for everything in the build at some
point and then move the verbose_subprocess_errors() context up to
main(), but this is the problem I'm debugging today and it'll do for
now.

Change-Id: Ie10b034ecdee653afa416947d38c0af2c61a2044