Make the fortify test more robust.
Which functions are covered by `_FORTIFY_SOURCE`` at compile time is
sort of an implementation detail. This test is supposed to check that
fortify is enabled by default, and we don't need to check specific
behavior to do this.

This test breaks with a newer sysroot because strcpy is no longer
covered by bionic's own diagnostics, but is instead covered by
`-Wfortify-source`, which is a separate set of checks from those enabled
by `_FORTIFY_SOURCE`.

Change-Id: Ib092eb9268db1528c92e2098712f6a8cecfed18c